Corrected stop codon frame (fix #39)
The frame must be computed separately for each feature.
The frames of the start codon and of the CDS are always the same, since
the start codon is part of the CDS.
The frame of the stop codon might differ if the stop codon is split over
two exons of if CDS has length that is not a multiple of three.